Separate OTPInputBlocks to other file
